fixed imports, install dependencies
